+++ src/ResponsiveTable/ResponsiveTable.js
@@ -1,5 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import TableHeaders from '../TableHeaders';
+import TableFooter from '../TableFooter';
 import TableRow from '../TableRow';
 
 class ResponsiveTable extends Component {
@@ -35,6 +36,23 @@ class ResponsiveTable extends Component {
         return <TableHeaders headers={ headers } />;
     }
 
+    renderTableFooter() {
+        const {
+            headers,
+            headerData,
+        } = this.props;
+
+        if (!headers || !headers.length) {
+          return null;
+        }
+
+        if (headers.findIndex((header) => 'footerContent' in header) === -1 ) {
+          return null;
+        }
+
+        return <TableFooter headers={ headerData || headers } />;
+    }
+
     renderTableBody() {
         const {
             data,
@@ -64,6 +82,7 @@ class ResponsiveTable extends Component {
             <table className="ffe-responsive-table">
                 {this.renderTableCaption()}
                 {this.renderTableHeaders()}
+                {this.renderTableFooter()}
                 {this.renderTableBody()}
             </table>
           );